[GEARPUMP-328] FetchThread fetchSleepMS from Int to Long
currently *KafkaConfig.FETCH_SLEEP_MS_CONFIG* is a INT parameter , move it into a LONG parameter is better .

Author: darionyaphet <darion.yaphet@gmail.com>

Closes #199 from darionyaphet/GEARPUMP-328.

+++ external/kafka/src/main/java/org/apache/gearpump/streaming/kafka/util/KafkaConfig.java
@@ -142,7 +142,7 @@ public class KafkaConfig extends AbstractConfig implements Serializable {
             ConfigDef.Importance.LOW,
             FETCH_THRESHOLD_DOC)
         .define(FETCH_SLEEP_MS_CONFIG,
-            ConfigDef.Type.INT,
+            ConfigDef.Type.LONG,
             100,
             ConfigDef.Range.atLeast(0),
             ConfigDef.Importance.LOW,
